+++ src/main/java/com/tcd/repository/TipoConteudoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tcd.model.TipoConteudo;
 
-public interface TipoConteudoRepository extends JpaRepository<TipoConteudo, Integer>{
+public interface TipoConteudoRepository extends JpaRepository<TipoConteudo, Long>{
 
 }
